+++ src/sluicer/mcp_answers.py
@@ -21,6 +21,7 @@ from typing_extensions import Required, TypedDict
 ErrorCode = Literal[
     "missing_extra",
     "refused_by_robots",
+    "refused_by_site",
     "refused_address",
     "fetch_failed",
     "too_large",
@@ -32,6 +33,7 @@ ErrorCode = Literal[
 PageErrorCode = Literal[
     "missing_extra",
     "refused_by_robots",
+    "refused_by_site",
     "refused_address",
     "fetch_failed",
     "too_large",